+++ CHANGES.md
@@ -1,6 +1,8 @@
 ## 0.2.0
 
-- Use Enumerator::Lazy for resources instead of plain array
+- Use `Enumerator::Lazy` for resources instead of plain array
+- Add time-range filtering methods to `ChangeList`, `ChangeDump`, `ChangeListIndex` and `ChangeDumpIndex`
+- Expose time attributes directly on `Augmented` objects so as to avoid having to fish around in the metadata
 
 ## 0.1.3